better user data fetching

+++ app/guestbook/form/actions.ts
@@ -1,6 +1,7 @@
 "use server";
 
 import filter from "@/lib/filter";
+import getProvider from "@/lib/getProvider";
 import prisma from "@/lib/prismadb";
 import type { Session } from "next-auth";
 
@@ -13,11 +14,20 @@ export async function submitMessage(
     throw new Error("Message contains profanity or user is blacklisted");
   }
 
+  if (
+    !session ||
+    !session.user ||
+    !session.user.email ||
+    !session.user.providerAccountId
+  ) {
+    throw new Error("No session found");
+  }
+
   const existing = await prisma.guestbookMessage.findFirst({
     where: {
       OR: {
-        email: session.user.email as string | undefined,
-        providerAccountId: session.user.providerAccountId as string | undefined,
+        email: session.user.email,
+        providerAccountId: session.user.providerAccountId,
       },
     },
     select: {
@@ -39,10 +49,17 @@ export async function submitMessage(
       },
     });
   } else {
+    const provider = await getProvider(session.user.providerAccountId);
+
+    if (!provider) {
+      throw new Error("No provider found");
+    }
+
     await prisma.guestbookMessage.create({
       data: {
-        email: session.user.email as string,
-        providerAccountId: session.user.providerAccountId as string,
+        email: session.user.email,
+        provider,
+        providerAccountId: session.user.providerAccountId,
         message: message,
       },
       select: {
@@ -53,11 +70,20 @@ export async function submitMessage(
 }
 
 export async function deleteMessage(session: Session) {
+  if (
+    !session ||
+    !session.user ||
+    !session.user.email ||
+    !session.user.providerAccountId
+  ) {
+    throw new Error("No session found");
+  }
+
   await prisma.guestbookMessage.deleteMany({
     where: {
       OR: {
-        email: session.user.email as string | undefined,
-        providerAccountId: session.user.providerAccountId as string | undefined,
+        email: session.user.email,
+        providerAccountId: session.user.providerAccountId,
       },
     },
   });
